refactor: move dns pkg
